add support for sandbox-tethered executables
Add a small hook to `raco setup` to make copies of all installed
executables where the copies start with a different addon directory
than the default user-specific path. Whether those executables
are created is based on a configuration in the addon directory
itself.

Although the same effect can be achived by setting PLTADDONDIR, a
tethered `racket`, `raco`, `drracket`, etc., can be easier to work
with and compose better with other programs, along the same lines as
separate in-place development builds from the Git repo.