experiment with revised optional-argument protocol
Switch from a pair of arguments for optionals --- the value and a
boolean indicating whether the value is meaningful --- to a single
argument with `unsafe-undefined` to mean that argument wasn't
supplied. In common cases where the default expression is an
immediate, the immediate can be used instead of `unsafe-undefined`,
and a callee-side test can be removed.