cs: add cache for JIT mode
When linklets are compiled in JIT mode and a procedure is to be
compiled on demand, consult a cache of compiled fragments (by default
"jit.sqlite" in the addons directory) and use an existing compiled
fragment or add to the cache after compiling.

This preliminary experiment suggests that the idea is workable.
Starting a program a second time is almost as fast as starting a
program compiled in non-JIT mode (i.e., directory to machine code).
Some refinements are needed: limiting the size of the JIT-fragment
cache, better contention handling, and better inlining of structure
operations in JIT mode (which is likely useful to cross-linklet
optimization in non-JIT mode, too).