expander: fix tracking of submodule base scopes
The bug was related to re-expansion of submodules that are implemented
as `(module* _name #f ....)`. The bug affects `module->namespace`,
which relies on knowing the set of scopes that correspond to the
submodule's body. the bug probably affects some other things that
similarly rely on a module's body scopes.

When you expand a module

 (module _name _initial-import
    ....)

then the set of scopes associated with `_initial-import` count as the
module's body scopes, except that two more are added for the "inner
edge" and "outer edge". Those two are added back to `_initial-import`
in the expanded form of the module. If you expand again, that means the
re-expansion starts with those two scopes added by the previous
expansion; since those to scopes are "everywhere" in the module body
from the previous expansion, so far, so good.

When you expand the submodule in

 (module ....
    (module* sub #f
       ...)
    ....)

then the submodule's scopes should start as the scopes on the `#f`,
which will normally be the scopes for the enclosing module's body. At
some point fairly early in the implementation of the new macro
expander, I decided to ignore the `#f` and just propagate the scopes
from the enclosing module's body as the starting point for the
submodule.

Besides being inconsistent with the v6.x expander, that choice means
that when fresh "inner edge" and "outer edge" scopes are allocated for
the submodule expansion, they're not usefully preserved for
re-expansion when they're added to the `#f`. (In fact, the
implementation didn't bother adding them to `#f` in the expansion.) So,
those scopes get lost, and re-expansion has the wrong idea for the
submodule's body scopes.

Why didn't this cause more problems? Because you don't normally need
the concept of a module's body scopes. Forms within the module body
have various scopes, and normally all that matters is how those forms
relate to each other. But it matters for `module->namespace`, because
forms evaluated in the module's namespace need to be treated as if they
were in the module all along.

The repair is obvious: don't treat `#f` as special in this respect.
That's what you see in the patch: the `enclosing-all-scopes-stx`
argument goes away for `expand-module`, and `all-scopes-s` is always
the same as `initial-require-s`. (I left both names defined, though,
because various operations are more directly tied to one notion or the
other.)

The new test uses `(eval-syntax (expand ...))`, which implies a
re-expansion, and then it uses `module->namespace`.