disable single-flonum reading by default
Add `single-flonum-available?` and `read-single-flonum`, where the
latter controls whethers numbers that have an "s" or "f" exponent
merker are parsed as single-flonums are normal flonums. The parameter
is disabled by default, which changes the meaning of most existing
code that has a literal number with "s" or "f", including `+inf.f` and
`+nan.f`.

The compiler constant-folds `single-flonum-available?` and
`real->single-flonum` on a flonum, so use a combination of those to
replace most uses of a single-flonum literal. Single-flonums within
quoted data are less convenient.