expander: avoid race condition on `syntax-e`
Lazy scope propagation concurrently in two threads had a (very
unlikely) race condition updating a syntax object. Ryan pointed out
this problem as part of the discussion for #3162.

Switch to using a single field for a syntax object's content and
propagations, so a CAS can be used to update the field. As a pleasant
side effect, this change tends to make syntax objects more compact.