Chez Scheme: toward parallel GC
This step adds parallelism, but isn't yet any faster than non-parallel
GC. Hopefully, it's on the path to effective parallel GC.

All allocation is now thread-local, which recovers a small bit of
performance that was lost when adding thread-local allocation
alongside global allocation.

Parallelism uses thread contexts created by Chez Scheme threads (which
correspond to Racket places and future-running threads), but it
creates its own threads to perform collection. The number of
collection-helper threads is limited to the number of active Chez
Scheme threads. Only the main "sweep" pass runs in parallel --- that
is, after roots have been traversed, and before weak references and
finalization are handled --- but that should be the bulk of collection
work. Also, memory-accounting collections and generation-0-to-1
collections always runs sequentially.

The limitations may explain the lack of speedup, but perhaps more
likely it's the way that helper threads stall when they hit a
conflict. Further work is needed to increase throughput in the
collector, but just getting this much right (i.e., synchronized
correctly) was a big hurdle, and it seems worth testing it while
making further refinements.