Chez Scheme GC: internal parallelism by messages instead of locks
Changes the internal parallelism strategy for the GC to record an
owner for each allocated segment of memory, and have the owner be
solely responsible for copying or marking objectsown of the segment.
When sweeping, a collecting thread handles references to objects that
it owns or that have been copied or marked already, and it asks
another collecting thread to resweep an object that refers to other
objects owned by that that thread. At worst, an object ends up being
swept by all collecting threads, one at a time, but that would be very
rare.

Thie approach seems likely toscale better than a lock-based approach,
even the one that used a lightweight, CAS-based lock and retried on
lockk failure. It appears to improve parallelism in `raco setup`
during the bytecode compilation phase, but it does not yet help much
with tthe document-running and -rendering phases.