cs: simplify and improve handling of literals
When Racket code contains a literal that cannot be serialized directly
by Chez Scheme (such as a keyword or an immutable string that should
be datum-interned), the old approach was to generate Scheme code to
construct the literal through a `let` binding. To handle paths
associated with procedures, however, Chez Scheme's `fasl-write` now
allow arbitrary values to be intercepted during fasl and passed back
in to `fasl-read`. Using that strategy for all Racket literals
simplifies the implementation, reduces compiled code, and makes
closures smaller. It also increases the number of relocations, but
the code savings mean that DrRacket's foorprint shrinks by about 1%.