Chez Schme: change struct predicate implementation
Instead of keeping a rectord types' ancestry in a vector from subtype
to supertype, keep the vector the other way around, and include a
record type at the end of its own vector. This order makes a more
direct test possible when checking for a known record type, especially
one without a supertype. Also, add `record-instance?`, which in unsafe
mode can skip the check that its first argument is a record.