Unix: restore all signal handlers after creating a subprocess fork
Save the ID of each signal that is modified, and restore all to
default in a child process after a `fork` and before an `execve`. Also
clear the signal mask, just in case.

Tracking all of the modified signals requires cooperation from several
subsystems, which often takes the form of a callback registered with
the subsystem before it might adjust a signal handler.

Closes #3609