rktio: restore all signal handlers after creating a subprocess fork
Set every signal handler to `SIG_DFL` in a child process after a
`fork` and before an `execve`. Also clear the signal mask, just in
case.

In principle, we need to know the range of possible signal handlers
for each platform, but it looks like iterating from 1 to SIGPROF
covers things well enough.

Closes #3609