Chez Scheme: improve multiplication and scheduling
Adjust kernel bignum operations to decrease the trap counter, and when
an operation would take too long as single kernel operation, bounce
bac out to Scheme through a kind of trampoline continuation.
Otherwise, a program that performs many big multiplcations or
divisions does not check for Ctl-C or swap threads often enough.

Also, improve multiplication with trailing 0s and raise the threshold
for using Karatsuba. The experimentally determined threshold (on an M1
Mac) matches the GMP default threshold, so that seems like a good
sign. Meanwhile, multiplying bignums with trailing 0s is common enough
to be worth a special case.