move `memq` and `memv` back to core
This commit partly reverts c305dba649 and 88fc9a979f, which was
discusssed in PR 366.

The original discussion was about allowing an extra argument to
`member` to pick the equality predicate. That change is still in place
in the same (private) Racket module.

The `memq` and `memv` functions are in Scheme for Racket CS and back
to being and in C for Racket BC. The old motivation for moving `memv`
and `memq` --- to get them out of C --- is mostly subsumed by CS
(granting that some C code got added back to BC meanwhile). The
advantage of make `memq` and `memv` more core is that the compiler can
do more with them, at least in CS.

When `memq` and `memv` were moved previously, they lost checking for
cyclic lists. That wasn't discussed and presumably wasn't on purpose;
check is restored here, including for `member`.

There's no significant performance change in CS, except in cases like
`(memq x '(a b c))` that the compiler unpacks into a combination of
`eq?`s.