raco demodolarize: adapt to CS
Really, `raco demod` is adapted here to work with any linklet-based VM
by compiling modules to machine-independent form, which is essentially
a wrapper around linklet S-expressions. The BC-specific implementation
remains in place, and it has the advantage of being able to work with
existing module compilations, while the implementation based on
machine-independent form must recompile all modules before attempting
to combine them (but that recompilation is easily cached).

Getting `raco demod` to work involved several incidental improvements:

 * make `racket/linklet` work with machine-independent forms;

 * fix schemify for linklets that have unexported definitions (which
   the expander never generates, but the demodularizer does); and

 * add `current-multi-compile-any` to expose CM's multi-target
   compilation mode programmatically.

The demodularizer could be a lot smarter to prune demodularized code
before sending it off to the compiler. Of course, the compiler should
be able to figure out that improvements itself, but sending a smaller
chunk of code to the compiler might avoid the hybrid
interpreter--compiler mode that is used for large linklets.

The demodularizer has a lot in common with the expander's flattener
that is used for bootstrapping, and a smarter demodularizer would have
even more in common. It would be nice to have two implementations
instead of one.