Chez Scheme: flip tagging offset
Instead of tagging an object by subtracting from its address, tag by
adding to its address.

The difference is very minor, but it's like more intutitive to new
developers, and it saves an addition when stripping the tag from a
pointer whose type is not known a priori (which happens in the GC more
than other places).