add binding-space support to `provide` and `require`
The existing `make-interned-syntax-introducer` function supports a
binding-space convention where a scope (keyed on a symbol name) can be
used consistently across modules to create bindings with that scope.

This commit adds more direct support for binding spaces to `require`
and `provide`. A `for-space` form allows a module to provide a binding
in a space, and then `require` will (by default) create a binding in
an importing context using the space's scope.

The changes to `require` and `provide` are mostly internal, but the
result of a function like `identfier-binding` might need to include
require/provide space information as well as require/provide phase
information. To minimize changes to existing code, space information
is combined with phase information in a single value, and the default
space (i.e., no extra space scope) is represented by just using the
phase value. In other words, the representation of import and export
information is unchanged unless binding spaces are specifically used.
Require and provide transformers similarly need to change to work with
binding spaces, but existing transformers continue to work as long as
they are not combined with binding-space operations.