expander: add glue syntax with `#%require` support
Glue syntax helps implement importing with heirarchical names, which
is a feature of a current Rhombus prototype. Building only half of
hierarchical naming into the expander is not very satisfying; it would
be nicer either to build it completely outside the expander or to have
simple and direct suport provided the expander. Building things
outside the expander proved too challenging, and it was going to
defeat the expander's bulk-binding mode for imports, which can greatly
improve the representation and performance of import handling.
Meanwhile, it's not clear how to build hierarchical naming more fully
into the expander without extending the core language and creating
substantial backward-compatibility issues.

This change side-steps most backward-compatibility problems by adding
support into `#%require`. Few tools attempt to inspect a `#%require`
expansion; notably, `#%module-begin` forms that inspect expansion
terms normally treat `#%require` as opaque. (In contrast, a new
module-level from would create more trouble.) The handful of tools
that do inspect `#%require` expansions will need to be updated.

The new pieces:

 * `make-glue-syntax`, `glue-syntax?`, and `glue-syntax-target`: a new
   primitive value. It's is somewhat like a rename transformer in that
   it's normally used at compile time and bound to an identifier
   `for-syntax`.

 * `(glue <id> <tag> <spec>)` subform in `#%provide`: imports like
   `<spec>`, but adds a fresh scope to the `<spec>` imports.
   Meanwhile, `<id>` is bound to glue syntax of the form `(<tag> <id>
   <post-id>)`, where `<post-id>` is `<id>` plus the fresh scope. In
   other words, using `<post-id>` provides access to the bindings
   imported with `<spec>`. The `<tag>` can be anything and potentially
   distinguishes uses of `glue` for different purposes.

   The binding of `<id>` is like one that could be written as
   `(define-syntax <id> (make-glue-syntax ....))` , but the `glue`
   form is able to bind `<id>` in every phases that is represented by
   imports from `<spec>` --- including the label phase.

 * `identifier-binding-glue-syntax`: like `syntax-local-value`
    composed with `glue-syntax-target`, but can be used at any time,
    and does not require instantiating the module that binds an
    identifier to glue syntax. The reuslting syntax is phase shifted
    in a way that is consistent with the module being instantiated at
    the relevant phase (if it were to be instantiated), even to the
    label phase. The "no instantiation" part here is particularly
    related to using label-phase bindings for tasks like documentation
    linking.

 * `identifier-distinct-binding`: not really tied to glue syntax, but
    a handy shortcut to checking whether a scope is relevant to
    finding a binding, instead of comparing the results of tro calls
    to `identifier-binding`.

Constraining glue syntax to syntax objects, as opposed to allowing a
generative record type, ensures that getting a glue syntax target from
an uninstantiated module makes sense. Binding information on a "tag"
in a glue-syntax target can distinguish different kinds of glue
syntax.

A language with hierarchical naming can use these pieces to represent
some (or all) hierarchical names with glue-syntax bindings, and then
it gets the expander's fancy support for representing bindings more
efficiently as well as a way to use the label phase.