add hooks to rewrite primitive error messages
This is only one piece in making Racket error messages more
configurable to different languages, focusing on the part where
primitives should compile the same (with the same speed) but with
error messages that use different names and different contract syntax.
Error messages can be customized through four new parameters:

 error-primitive-name->symbol-handler - can substitute the symbol name
  of a Racket primitive with a different symbol name

 error-primitive-contract->string-handler - can rewrite the string that
  represents a contract in Racket syntax

 error-primitive-message->string-handler - can rewrite the text of a
  primitive error message; this will typically be a last resort, but
  it can be used to change "procedure application" to "function call"
  or chage "vector" to "array", for example

 error-struct-operation-names-handler - selects how a structure-type
  accessor name and mutator name is formed, along with the contract
  string representing the structure type (as the first argument to
  an accessor or mutator)

Using parameters has the benefit that they only slow down error
reporting and have no effect on compiled code. Parameters have the
drawback that they're not inherently sensitive to the context where a
primitive was misused, but instead to a global configuration that is
typically applied at startup. This is consistent with value printing
(i.e., value-printing configuration currently has the same drawback).