Chez Scheme: add pbchunks support
As of comment 4480e643da, Chez Scheme can support all of Racket CS on
a platform where there's no native-code backend or where code
generation more than libffi is disallowed. Then again, since
interpreted mode is a factor of 20 slower, and since that covers
everything from Scheme primitives to the compiler to the expander,
it's not so practical.

This commit makes a no-native-backed Racket CS somewhat more practical
by adding a "pbchunks" mode that compiles bytecode to C, at least to
some degree. That compilation path makes sense for the static code
that is built into the Racket CS executables, which includes the
layers from Chez Scheme primitive through the Racker expander. This
strategy is analogous to the "cify" mode of non-JIT Racket BC, which
compiles the Racket macro expander to C.

One different between pbchunks and cify is that Racket BC's primitives
are already implemented in C, so cify applies only to the expander's
implementation, but pbchunks applies even to the implementation of
Scheme primitives. (Neither applies to Racket-level code that is run
via no-JIT Racket BC with cify or no-native-backend Racket CS with
pbchunks.) That difference explains the relative performance I find on
my machines for `racket -cl racket/base`

                                 pure
                 native       interpreted       cify/pbchunks
  Racket CS       x1             x20                x4
  Racket BC       x1             x20                x1

This x4 is not great, of course, but the end result is that building a
Racket distirbution with pbchunks Racket CS feels at least practical,
while pure interpreted Racket CS took longer than I was willing to
wait.