Chez Scheme: repairs for misaligned word access
On a platform that requires aligned access for 64-bit values (e.g.,
sparc64), return-address information that is embedded in code is not
necessary 8-byte-aligned, since instructions are each 4 bytes.
Currently, there are no native-code backends with this problem, but
it's a problem for 64-bit pb. When accessing return-address
information, Use load and store operations that support unaligned
access.