unzip: add `#:preserve-attributes?`
Support applying file permissions when unpacking files. Enabling this
mode turns out to imply several changes.

One change is to use the archive's central directory to find files,
because permissions attributes are only stored there, and not in the
per-file header that duplicates most other information. In fact, while
the old implementation of `unzip` scans through the file to read all
entries, that's not the way zip archives are supposed to work. Files
in an archive might be "deleted" without removing their data and just
updating the central directory at the end the archive, and those files
should not be unpacked. To support ports that allow only streaming and
not `file-position`, however, the old approach is still used when the
`#:preserve-attributes?` argument is `#f`.

A second change is to extend the protocol for entry readers. A prvious
change was made to support `#:preserve-timestamps?`, but it wasn't
general enough, and we double down on the approach to extension
without breaking old code. Besides changing an argument to an entry
reader, `#:preserve-attributes?` mode also changes the result: a
reader can return a post thunk that is run after all extraction
operations, which is needed for setting directory permissions after
files are written to the directory, for example.