unzip: add `#:preserve-attributes?`
Add support for perserving file permissions in addition to just
timestamps.

This time, the expanded interface can better accomodate future
extensions. The problem last time was likely that information like
permissions did not appear to be available, but that's because it's
only in the central directory instead of the redundant file header.
There still doesn't appear to be anything else useful to pass along,
but it won't be so bad when it turns out that there's something else
that I'm overlooking.