Chez Scheme: adjust default collect-request handler
Instead of always performing a major (i.e., maximum generation)
collection on a schedule, which is every 64th collection by default,
waiting until the amount of memory grows by a factor (2 by default).
This change should be a better default for applications that have a
lot of long-lived data and also allocate a lot of short-lived data,
which is a common pattern.