CS: static FFI call generation
The `ffi/unsafe` API is an especialy dynamic interface, which made
sense for BC where all code generation is dynamic. For
direct-to-machine-code Chez Scheme, `ffi/unsafe` has been implemented
using `eval` on demand, but that pushes work to run time. Possibly
more important in the long run, dynamic code generation isn't
compatible with a possible future build mode that omits the Chez
Scheme compiler.

Instead of changing the `ffi/unsafe` API, this commit approaches
static generation as an optimization problem. The optimization is
sufficient right now for all bidnings the minimal Racket and in the
"draw-lib" package, except for `net/macosx-ssl` (which needs support
for struct C types).

The optimization is a cooperative effort in four parts:

 * The Rumble layer provides a new form
   `ffi-static-call-and-callback-core`, which generates the core part
   of a foreign callout or callback at compile time based on Chez
   Scheme foreign representations. The Rumble layer also provides
   `ffi-maybe-call-and-callback-core`, which doesn't do anything, but
   is a hint to the schemify layer. Finally, the Rumble layer provides
   `assert-ctype-representation`, which is checks its arguments only
   when the Rumble layer is compiled in safe mode.

 * The `unsafe/ffi` layer expands `_fun` forms to use
   `ffi-maybe-call-and-callback-core` as a hint to the schemify layer.
   It also takes care to define C types in a way that is visible to
   schemify's constant/copy propagation, including using
   `assert-ctype-representation` as needed. A few functions exported
   from `unsafe/ffi` have turned into macros that add an aseertion
   wrapper.

 * The schemify layer recognizes the primitive C type definitions and
   maps them to Chez Scheme representations, propagates representation
   information through definitions and across module boundaries, and
   detects when a `ffi-maybe-call-and-callback-core` can be converted
   to `ffi-static-call-and-callback-core` based on that propoagated
   information.

 * The cross-compilation protocol adds support queries that are
   written as dynamic choices for C type representations, but constant
   folding and propagation can turn them into compile-time choices.
   The schemify layer uses this compile-time query interface to enable
   compile-time propagation of C type representations.

The optimization path does not yet handle struct and union arguments
or results (as opposed to a pointer to a struct or union, which is
handled).

The new `ffi/unsafe/static` library exports the same binindings as
`ffi/unsafe`, except that it replaces `_fun` with one that will
trigger a compile-time error if a `ffi-maybe-call-and-callback-core`
cannot be converted into `ffi-static-call-and-callback-core`. The
intent is that libraries use `ffi/unsafe/static` to make sure that
dynamic code generation will not be needed by the module. There's a
danger that some future schemify revision loses constant-propagation
power and fails to compile a module using `ffi/unsafe/static`, but the
idea is that such a regression should be treated as a bug, and
hopefully `ffi/unsafe/static` will help catch it.

For BC, `ffi-maybe-call-and-callback-core` simply returns its
arguments as in CS, and there is no conversion to a form for static
code generation.

The performance difference from static code generation tends to be
very small, because on-demand compilation of foreign callouts and
callbacks is already lazy, cached, and fast, and because the core of a
callback is always wrapped further to complete checking and
thread-cooperation details. The long-term goal is mainly to have a
path to compilation-free run time, although further improvements may
one ate produce performance benefits. Meanwhile, improvements that
were needed in schemify's optimizations might provide small benefits
independent of foreign calls.