CS: static FFI call generation
The `ffi/unsafe` API is an especialy dynamic interface, which made
sense for BC where all code generation is dynamic. For
direct-to-machine-code Chez Scheme, `ffi/unsafe` has been implemented
using `eval` on demand, but that pushes work to run time and, in the
long run, isn't compatible with a build mode that leaves out the
compiler.

Instead of changing the `ffi/unsafe` API, this commit approaches
static generation as an optimization problem. It works for bindings in
the "openssl" collection, "net" collection, and the "draw-lib"
package.

The optimization is a coopeartive effort in four parts:

 * The rumble layer provides a new form
   `ffi-static-call-and-callback-core`, which generates the core part
   of a foreign callout or callback at compile time based on
   representations at the Chez Scheme level. The Rumble also provides
   `ffi-maybe-call-and-callback-core`, which in principle does the
   same sort of thing dynamically, but actually always declaines to
   produce core implementation and leaves things to to existing path.
   Finally, the Rumble layer provides `assert-ctype-representation`,
   which is checked only when the Rumble layer is compiled in safe
   mode.

 * The `unsafe/ffi` layer, which expands `_fun` forms to use
   `ffi-maybe-call-and-callback-core` as a hint to the schemify layer.
   It also takes cre to define C types in a way that is visible to
   schemify's constant/copy propagation, including using
   `assert-ctype-representation` as needed. A few function exported
   from `unsafe/ffi` have turned into macros that add an aseertion
   wrapper.

 * The schemify layer, which recognizes the primitive C type
   definitions and maps them to Chez Scheme representations,
   propagates representation information through definitions and
   acorss module boundaries, and detects when a
   `ffi-maybe-call-and-callback-core` can be converted to
   `ffi-static-call-and-callback-core` based on that propoagated
   information.

 * The cross-compilation protocol, which needs to support queries that
   are written as dynamic choices for C type representations, but
   through constant folding and propagation can turn into compile-time
   choices (to enable compile-time propagating of C type
   representations).

The optimization path does not yet handle struct and union arguments
or results (as opposed to a pointer to a struct or union, which is
handled here).

For BC, `ffi-maybe-call-and-callback-core` simply returns its
arguments as in CS, and there is no conversion to a form for static
code generation.

The performance diference tends to be very small, because on-demand
compilation of foreign callouts and callbacks is already lazy, cached,
and fast, and because the core of a callback is wrapped further to
complete checking and thread-cooperation details. The long-term goal
is mainly to have a path to compilation-free run time, although
further changes may accrue some performance benefits.

Necessary improvements to schemify's optimizations might provide small
benefits independent of foreign libraries.